refactored hook names, removed applicationHistory service
