--- src/hooks/create-course-input.ts
@@ -1,15 +1,0 @@
-/**
- * @file create-course-input.ts - create course input hook
- * @author cmc
- * @since v0.0.1
- */
-import { Hook, HookContext } from '@feathersjs/feathers'
-
-// eslint-disable-next-line @typescript-eslint/no-unused-vars
-export default (options = {}): Hook => {
-  return async (context: HookContext): Promise<HookContext> => {
-    console.log(context.result)
-
-    return context
-  }
-}
