--- src/services/author-applications/application-history/application-history.class.ts
@@ -1,9 +1,0 @@
-import { Service, SequelizeServiceOptions } from 'feathers-sequelize'
-import { Application } from '../../../declarations'
-
-export class ApplicationHistory extends Service {
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  constructor (options: Partial<SequelizeServiceOptions>, app: Application) {
-    super(options)
-  }
-}
