+++ src/services/courses/course-content/course-content.hooks.ts
@@ -1,6 +1,6 @@
 // import { HooksObject } from '@feathersjs/feathers'
 import * as authentication from '@feathersjs/authentication'
-import setCourseInput from '../../../hooks/set-course-input'
+import setCourseInput from '../../../hooks/course-content-input-set'
 // Don't remove this comment. It's needed to format import lines nicely.
 
 const { authenticate } = authentication.hooks
